pyproject.toml
[project]
name = "vta"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.14.1",
    "firebase-admin>=7.1.0",
    "google-genai>=1.56.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "resend>=2.19.0",
]
